fix output callback kwarg edge cases
If `save_start` was false, `u_modified` was not set to false for SavingCallback and FunctionCallingCallback. This led to a doubly saved t=0, but only when combined in a CallbackSet.

Furthermore we now properly support combining scalar saveat from #157 with excluding the start or end.

Test cases for both issues are included.