remove complication from PresetTimeCallback
Previously this code was doing some fairly weird stuff with the closures relying on indices being passed around in spooky ways. This was (I believe) intended as an optimization, but was confusing both me and the compiler, so I deleted it, and am replacing the in calls with `insorted` to make it faster.